autopilot: move handlers back to autopilot.go
